% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_slct.R
\name{bar_slct}
\alias{bar_slct}
\title{Extract data.frame and barplot}
\usage{
bar_slct(dat, mrgn, slct, target, pm, nm, yscale, gg, wrapcol, xlb, ylb, ...)
}
\arguments{
\item{dat}{data.frame}

\item{mrgn}{integer 1 or 2 search row is 1, search col is 2}

\item{slct}{character}

\item{target}{character}

\item{pm}{character strings. A type of pattern match, as 'mgrep', 'mmatch', or 'match'.
The default value is 'mmatch'}

\item{nm}{character vector, as name of labels of barplot.
If it is 'NULL', names or rownames of data.frame will used as 'nm'.}

\item{yscale}{character 'free' or 'fixed'. The default value is 'free'}

\item{gg}{logical default value is FALSE}

\item{wrapcol}{integer}

\item{xlb}{character}

\item{ylb}{character}

\item{...}{other barplot options}
}
\description{
Extract data.frame and barplot
}
\examples{
\dontrun{
# sample data
zdat <- rsko::zconv(iris, -5, 1)
nfpkm <- rskodat::nfpkm[1:36,]
kodat <- rskodat::kodat
cpdat <- rskodat::cpdat
cge100 <- rskodat::cge100

# search strings
slct1 <- c('1', '51', '101')
slct2 <- c('gene1', 'gene10', 'gene150')
slct3 <- c('K07230', 'K02030', 'K02073')
slct4 <- c('100770489', '100689463', '100774853')

# subset rows of data.frame
par(mar=c(6, 4, 2, 1), mfrow=c(1,3))
res1 <- rsko::bar_slct(zdat[-5], mrgn=1, slct=slct1, target=rownames(zdat), yscale='fixed')
res2 <- rsko::bar_slct(zdat[-5], mrgn=1, slct=slct1, target=rownames(zdat), yscale='fixed', gg=TRUE)
print(res2)

# subset columns of data.frame
res3 <- rsko::bar_slct(dat=nfpkm[-1:-4], mrgn=2, slct=slct2,
 target=names(nfpkm[-1:-4]), nm=nfpkm$id)
res4 <- rsko::bar_slct(dat=nfpkm[-1:-4], mrgn=2, slct=slct2,
 target=names(nfpkm[-1:-4]), gg=TRUE, xlb = 'sample', ylb = 'fpkm')

# subset rows of data.frame
par(mfrow=c(1,3))
res4 <- rsko::bar_slct(dat=kodat, mrgn=1, slct=slct3, target=rownames(kodat))
res5 <- rsko::bar_slct(dat=kodat, mrgn=1, slct=slct3, target=rownames(kodat), yscale='fixed')
res6 <- rsko::bar_slct(dat=kodat, mrgn=1, slct=slct3, target=rownames(kodat), gg=TRUE,
                 yscale = 'fixed', wrapcol=3, xlb='condition', ylb='relative exp.')

# partial match of search strings　
res7 <- rsko::bar_slct(dat=cge100, mrgn=1, slct=slct4, target=rownames(cge100),
                 pm = 'mgrep', yscale = 'fixed')

# result of degview genes
res.degv <- rskodat::res.degv
res8 <- lapply(1:nrow(res.degv), function(i){
                  subg <- unlist(strsplit(unlist(res.degv[i, 'kos']), ';'))
                  rsko::bar_slct(dat=kodat, mrgn=1, slct=subg, yscale = 'fixed',
                  target=rownames(kodat), gg=TRUE)
                  })
# result of degview cpds
res9 <- lapply(1:nrow(res.degv), function(i){
                  subg <- unlist(strsplit(unlist(res.degv[i, 'cpds']), ';'))
                  rsko::bar_slct(dat=cpdat, mrgn=1, slct=subg, target=rownames(cpdat),
                                 yscale = 'fixed', gg=TRUE)
                  })
}
}
