% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindfill.R
\name{bindfill}
\alias{bindfill}
\title{bindfill}
\usage{
bindfill(x, margin, fill)
}
\arguments{
\item{x}{target list}

\item{margin}{character: "c" as column or "r" as row.}

\item{fill}{fill strings: default value is NA}
}
\value{
list
}
\description{
bindfill
}
\examples{
#
v <- list(letters[1:4], letters[5:7], letters[8:16])
res1 <- rsko::bindfill(v)
res2 <- rsko::bindfill(v, fill = "")
}
