% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors657.R
\name{colours657}
\alias{colours657}
\title{colours657 window}
\usage{
colours657(distinct)
}
\arguments{
\item{distinct}{logical an argument for colours function, the default value is TRUE}
}
\value{
ggplot object
}
\description{
colours657 window
}
\examples{
#
res <- colours657()
}
