% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corheat.R
\name{corheat}
\alias{corheat}
\title{Gene clustering and cutting cluster with dynamic cut}
\usage{
corheat(dat, distm, clm, method_dycut, draw)
}
\arguments{
\item{dat}{RNA-seq count data(row:sample, column:gene)}

\item{distm}{select from c("pearson", "abspearson", "squarepearson", "spearman")}

\item{clm}{clustering method of hclust}

\item{method_dycut}{method of 'cutreeDynamic' select from "tree", or "hybrid}

\item{draw}{logical draw heatmap or nod.}
}
\value{
draw heatmap and results of dynamic cut
}
\description{
Gene clustering for RNA-seq data and cutting cluster with dynamic cut
}
\examples{
\dontrun{
# data: normalized fpkm
nfpkm <- as.data.frame(rskodat::nfpkm)
res.dcut <-cornet::dycutdf(nfpkm[-1:-4], distm = "correlation")
cl_dat <- res.dcut$cluster_dat
corheat with dynamic tree cut
res <- corheat(dat = nfpkm[-1:-4], distm = "correlation", clm = "average",
               method_dycut = "tree", draw = TRUE)
}
}
