% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degview.R
\name{degview}
\alias{degview}
\title{Mapping of gene and/or compound expression data on KEGG pathway map using pathview.}
\usage{
degview(gene_dat, cpd_dat, pathway_id, org, outdir, ...)
}
\arguments{
\item{gene_dat}{named numeric vector, matrix, or data.frame. The 'gene_dat' has
row.names or which is kegg gene id(cge:100770489) or kegg orthology id(K01488).
The organism code is removed automatically(100770489).}

\item{cpd_dat}{data.frame has row.names which is compoud id.  e.g. C00631}

\item{pathway_id}{pathway id E.g. c("00010", "04210"). If it is "all", all pathways which
genes and/or compunds were mapped.}

\item{org}{organism code E.g. "hsa"(human), or "ko"}

\item{outdir}{the directory of KEGG pathway data file (.xml) and image file (.png).
result of image files (.multi.png) created at current working directory.}

\item{...}{other pathwiew options}
}
\value{
the nested list. The result of pathview and data.frame of ids per map.
}
\description{
Mapping transcriptome and/or metabolome data on KEGG pathway
}
\examples{
\dontrun{
# arguments: gene_dat and cpd_dat
cpdat <- rskodat::cpdat
kodat <- rskodat::kodat
cge100 <- rskodat::cge100

# arguments: pathway_id
maps1 <- c("00230", "00240")

# execute:
# both gene and compound
suppressPackageStartupMessages(library(pathview))
res1 <- degview(gene_dat=kodat, cpd_dat=cpdat, pathway_id=maps1, org="ko",
                outdir="~/pub/tmp")
# gene:
res2 <- degview(gene_dat=kodat, pathway_id=maps1, org="ko", outdir="~/pub/tmp")

# compound:
res3 <- degview(cpd_dat=cpdat, pathway_id=maps1, org="ko", outdir="~/pub/tmp")

# organism code is "cge":
res4 <- degview(cge100, pathway_id = maps1, org = "cge", outdir = "~/pub/tmp")

# result:
head(res1$id_table); str(res1$gene[1:3]); str(res1$compound[1:3])

# result of pathview:
head(res1$`00230`$plot.data.gene)

# additional
tibble::rownames_to_column(res1$gene$`00230`, "gene") \%>\%
   tidyr::gather(., key="key", value="v", -1) \%>\%
   ggplot2::ggplot(., ggplot2::aes(x=key, y=v)) +
       ggplot2::geom_bar(stat = "identity") +
       ggplot2::facet_wrap(~gene, ncol=3)
}
}
