% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utls.R
\name{df_utls}
\alias{df_utls}
\alias{td}
\alias{subd}
\title{The utility for handling data.frame}
\usage{
td(dat, nm, fctr_dat)

subd(dat, idlist, idx, mrgn)
}
\arguments{
\item{dat}{data.frame}

\item{nm}{character}

\item{fctr_dat}{data.frame or vector}

\item{idlist}{list: a list for extracting subset}

\item{idx}{atomic: a column of index for splitting the data.frame}

\item{mrgn}{integer: margin of data.frame}
}
\description{
The utility for handling data.frame
}
\examples{
\dontrun{
# Create a transposed data.frame with names and columns of factor
dat <- rskodat::fpkm
nm <- rownames(dat)
fctr_dat <- rsko::splt_dat(names(dat), "_", c("cnd","time","rep"))
tdat <- td(dat, nm, fctr_dat)
tdat[1:6]

# Subset with a list for index column
slst <- list(toyota = c("Toyota Corolla", "Toyota Corona"),
    mazda = c("Mazda RX4", "Mazda RX4 Wag"))
subd(mtcars, slst, rownames(mtcars), 1)
}
}
