% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utls.R
\name{fct_utls}
\alias{fct_utls}
\alias{fct_add}
\title{The utility of handling factors}
\usage{
fct_add(x, v)
}
\arguments{
\item{x}{factor}

\item{v}{atomic}
}
\description{
The utility for handling factors
}
\examples{
# unique length
x <- factor(c(1,3,5))
v <- 1:10
fct_add(x, v)
}
