% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcols.R
\name{ggcols}
\alias{ggcols}
\title{Create colour code using ggplot2}
\usage{
ggcols(n)
}
\arguments{
\item{n}{numbers of colour code}
}
\description{
Create defaoult colour code of ggplot2
}
\examples{
## ggcols
cols <- ggcols(n = 10)
barplot(setNames(rep(1,10), cols), col = cols, las =2)
}
