% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggutils.R
\name{ggutils}
\alias{ggutils}
\alias{gg_leg}
\alias{unfacet}
\alias{splitFacet}
\title{My utility functions for handling ggplot objects.}
\usage{
gg_leg(x)

unfacet(x)

splitFacet(x)
}
\arguments{
\item{x}{ggplot object}
}
\description{
My utility functions for handling ggplot objects.
}
\examples{
\dontrun{
# gg_leg
gg1 <- ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Sepal.Width, colour = Species)) +
 ggplot2::geom_point()
gg2 <- ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Petal.Length, colour = Species)) +
 ggplot2::geom_point()

res1 <- gg_leg(gg1); res2 <- gg_leg(gg2)
gg1_nl <- res1[[1]]; gg2_nl <- res2[[1]]; lgnd <- res1[[2]]

gridExtra::grid.arrange(gg1, gg2, ncol=2)
gridExtra::grid.arrange(gridExtra::arrangeGrob(gg1_nl, gg2_nl, lgnd,
 ncol = 3, widths = c(3/7, 3/7, 1/7)))

# splitFacet
gg <- ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Petal.Length)) +
 ggplot2::geom_point() +
 ggplot2::facet_wrap(~Species)
lst <- splitFacet(gg)
do.call(gridExtra::grid.arrange, c(lst, list(ncol = 3)))

}

}
