% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtable.R
\name{mdtable}
\alias{mdtable}
\title{Convert data.frame to table for markdown format}
\usage{
mdtable(dat, add_row)
}
\arguments{
\item{dat}{data.frame or matrix}

\item{add_row}{logical: add rownames to column or not.}
}
\description{
Mapping transcriptome and/or metabolome data on KEGG pathway
}
\examples{
\dontrun{ ##
dfiris <- head(iris); tbliris <- dplyr::as.tbl(head(iris)); mtxiris <- as.matrix(head(iris))
mdtable(dat = dfiris, add_row = TRUE)
mdtable(dat = tbliris)
mdtable(dat = mtxiris)
}
}
