% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstrings.R
\name{mstrings}
\alias{mstrings}
\alias{mgrep}
\alias{mmatch}
\alias{mgrep_ext}
\alias{mmatch_ext}
\title{Multiple patterns or values matching}
\usage{
mgrep(patterns, x, na.rm, value)

mmatch(patterns, x, na.rm = FALSE)

mgrep_ext(patterns, x, y = x, na.rm = TRUE)

mmatch_ext(patterns, x, y, na.rm = TRUE)
}
\arguments{
\item{patterns}{atomic}

\item{x}{atomic}

\item{na.rm}{logical}

\item{value}{logical}

\item{y}{atomic}
}
\description{
Multiple patterns or values matching
}
\examples{
\dontrun{
## Multiple patterns matching
x <- c("A,D", "E,F", "G,C", "C,H")
p1 <- c("A", "B", "C", "D")
p2 <- c("X", "Y")
p3 <- "A"

mgrep(p1, x)
mgrep(p1, x, F, T)
mgrep(p1, x, T, F)
mgrep(p1, x, T, T)

## If none of the patterns match, a list consisting of NAs or an empty list is returned.
mgrep(p2, x)
mgrep(p2, x, na.rm = T)

## Return as a list even if there is only one pattern
mgrep(p3, x)

## Multiple values matching
p4 <- c("A", "B", "C", "D", "X")
p5 <- c("H", "J", "K")
v2 <- c("A", "B", "C", "D", "B", "C", "D", "A", "B", "C")
mmatch(p4, v2)
mmatch(p4, v2, na.rm = T)
mmatch(p5, v2)
mmatch(p5, v2, na.rm = T)

## Multiple values mathing and extract correspondig data.
p <- c("A", "B", "C", "D")
x <- c("A,D", "E,F", "G,C", "C,H")
y <- c("v1", "v2", "v3", "v4")
dat <- data.frame(x, y, stringsAsFactors = FALSE)

mgrep_ext(p, dat$x, dat$y)

## Multiple values matching and extract corresponding data.
p <- c("A", "B", "C", "D", "X")
x <- c("A", "B", "C", "D", "B", "C", "D", "A", "B", "C")
y <- rpois(10, 5)
dat <- data.frame(x, y, stringsAsFactors = FALSE)

mmatch_ext(p, x, y)
}

}
