% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvenn.R
\name{mvenn}
\alias{mvenn}
\title{Multiple venn diagram, between all combinations of two list elements.}
\usage{
mvenn(x, y, show.plt, ...)
}
\arguments{
\item{x}{list}

\item{y}{list [default = NULL]}

\item{show.plt}{logical: draw plot or not. [default=F]}

\item{...}{other gplots::venn options}
}
\description{
Multiple venn diagram, between all combinations of two list elements. If only one list,
all combinations in the list elements, using the function gplots::venn.
}
\examples{
\dontrun{
# sample data
set.seed(1)
n <- 3; n_1 <- c(5,6,4); n_2 <- c(4,6,4)
v1 <- setNames(lapply(1:n, function(i){
 sample(LETTERS[1:10], n_1[i])
}), paste0("e",as.character(1:n)) )

v2 <- setNames(lapply(1:n, function(i){
 sample(LETTERS[1:10], n_2[i])
}) , paste0("f",as.character(1:n)))

# venn
mvenn(x = v1, show.plt = T)
mvenn(x = v1, y = v2)
}

}
