% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcat.R
\name{pcat}
\alias{pcat}
\title{PCA with prcomp and plot of select pc score}
\usage{
pcat(dat, x, y, scaling, cols, txt, tpos, ...)
}
\arguments{
\item{dat}{dataframe, matrix or prcomp object}

\item{x, y}{axis of pca score. 'x=1, y=2' means xaxis for PC1 score and yaxis for PC2 score}

\item{scaling}{prcomp option, scaling or not. The default is "TRUE"}

\item{cols}{A vector of manual colour}

\item{txt}{A size of text(optional)}

\item{tpos}{text position(optional)}

\item{...}{Further plot arguments. E.g., pch}
}
\value{
return prcomp object
}
\description{
This functions returns a result of prcomp object.
}
\examples{
# minimum usage
res <- pcat(iris[-5], scaling = TRUE, cols = c(1:3)[iris$Species])
res <- pcat(iris[-5], x =1, y = 3, scaling = TRUE, cols = c(1:3)[iris$Species], pch = 20)
}
