% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcfqs.R
\name{prcfq}
\alias{prcfq}
\alias{prcfqs}
\alias{prcfqlen}
\alias{prcfqbox}
\title{Process fastq file}
\usage{
prcfqs(filepath, n)

prcfqlen(filepath)

prcfqbox(filepath, n, ...)

prcfqlen(filepath)

prcfqbox(filepath, n, ...)
}
\arguments{
\item{filepath}{character}

\item{n}{Size of sampling [default: 10000]}

\item{...}{additional plot options [required for `prcfqbox`]}
}
\value{
data frame / integer vector
}
\description{
Process fastq file and create a data.frame of quality score
}
\examples{
\dontrun{
 in_f <- "test.fastq.bz2"
 res <- prcfqs(in_f)


}

}
