% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{PubMed search}
\usage{
pubmed_search(search_str, years, outfmt)
}
\arguments{
\item{search_str}{character: search strings using pubmed filter. E.g. [ti],[tiab],[tw]
\href{https://www.ncbi.nlm.nih.gov/books/NBK269313/}{Literature Search Methods}}

\item{years}{integer vector: E.g. 1980:2017}

\item{outfmt}{character: "tab" is a table of results ,or "cnt" is a count of hits per years. The default value is "tab".}
}
\value{
dataframe
}
\description{
This functions returns dataframe, several .
}
\examples{
## pubmed table
\dontrun{
str1 <- "microarray[tiab]"
str2 <- "next-generation sequencing[tiab]"
str3 <- "microarray[tiab] AND next-generation sequencing[tiab]"
str4 <- "cpa2[gs]"
str4 <- "MinION[tiab]"
yr <- 1980:2018

res1 <- pubmed_search(search_str = str1, years = yr, outfmt = "tab")
res2 <- pubmed_search(search_str = str2, years = yr, outfmt = "tab")
res3 <- pubmed_search(search_str = str3, years = yr, outfmt = "tab")
res4 <- pubmed_search(search_str = str4, years = yr, outfmt = "tab")

## count
cnt1 <- pubmed_search(search_str = str1, years = yr, outfmt = "cnt")
cnt2 <- pubmed_search(search_str = str2, years = yr, outfmt = "cnt")

## count plot
barplot(cnt1, las=2, ylim=c(0,6000), col=adjustcolor(2, 0.5), border = adjustcolor(2, 0.5));
par(new=T);
barplot(cnt2, las=2, ylim=c(0,6000), col=adjustcolor(4, 0.5), border = adjustcolor(4, 0.5))
legend("topleft",
       legend = c("microarray[title & abstract]",
                  "next-generation sequencing[title & abstract]"),
       col = adjustcolor(c(2,4), 0.5), pch = 15, bg = "transparent", box.lwd = NA)
}
}
