% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw.R
\name{slw}
\alias{slw}
\title{Sliding window for vector elements}
\usage{
slw(x, w)
}
\arguments{
\item{x}{target vector}

\item{w}{integer: window size}
}
\value{
list
}
\description{
Sliding window for vector elements
}
\examples{
#
v1 <- 1:13
v2 <- letters[1:13]
idx1 <- slw(v1, 3)
idx2 <- slw(v2, 3)
}
