% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectro_ndp.R
\name{spectro_ndp}
\alias{spectro_ndp}
\title{parsing a result file of nanodrop}
\usage{
spectro_ndp(fp, smp, facet_col)
}
\arguments{
\item{fp}{nanodrop result file path}

\item{smp}{character: sample names}

\item{facet_col}{integer: default value is 0L}
}
\value{
data.frame
}
\description{
A tsv file of nanodrop output.
}
\examples{
# arguments
fp1 <- system.file("extdata/nanodrop_out1.tsv", package="rskodat")
smp1 <- c("YL_N_1","YL_N_2", "ML_N_1","ML_N_2")

# execution
res1 <- spectro_ndp(fp1, smp1, 2L)
# res2 <- spectro_ndp(fp1, NULL, 2L)
# res3 <- spectro_ndp(fp1, NULL, 0L)
do.call(gridExtra::grid.arrange, c(res1$gg_list, list(ncol = 4)))

}
