% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_num.R
\name{str_num}
\alias{str_num}
\title{Count the number of matches in a list elements.}
\usage{
str_num(query, target, rate)
}
\arguments{
\item{query}{list}

\item{target}{list}

\item{rate}{logical}
}
\description{
Count the number of matches between all combinations of two list elements
}
\examples{
\dontrun{
set.seed(1)
l_target <- lapply(1:10, function(i) sample(letters, 5)) \%>\%
    setNames(., paste0("t",sprintf("\%02d", 1:10)))
l_search <- lapply(1:3, function(i) sample(letters, 3)) \%>\%
    setNames(., paste0("s",sprintf("\%02d", 1:3)))

res <- str_num(query = l_search, target = l_target, rate = T)
}
}
