% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{td}
\alias{td}
\title{Create a transposed data.frame with names and columns of factor}
\usage{
td(dat, nm, fctr_dat)
}
\arguments{
\item{dat}{data.frame}

\item{nm}{character}

\item{fctr_dat}{data.frame or vector}
}
\description{
Create a transposed data.frame with names and columns of factor
}
\examples{
\dontrun{
dat <- rskodat::fpkm
nm <- rownames(dat)
fctr_dat <- rsko::splt_dat(names(dat), "_", c("cnd","time","rep"))
tdat <- td(dat, nm, fctr_dat)
tdat[1:6]
}
}
