% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{utls}
\alias{utls}
\alias{ul}
\alias{dup}
\alias{char_sort}
\title{My R utility functions.}
\usage{
ul(x)

dup(x)

char_sort(x)
}
\arguments{
\item{x}{atomic}
}
\description{
My R utility functions.
}
\examples{
# return the list of position of duplicated value.
set.seed(12)
v1 <- sample(letters[1:5], size = 5, TRUE)
v2 <- sample(letters[1:5], size = 5, FALSE)
v3 <- sort(paste0("V", as.character(1:13)))
ul(v1)
dup(v1)
dup(v2)
char_sort(v3)

}
