% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcomp.R
\name{vcomp}
\alias{vcomp}
\title{Compares the elements of two lists against each other and outputs a count table}
\usage{
vcomp(query, target, rate)
}
\arguments{
\item{query}{list}

\item{target}{list}

\item{rate}{logical}
}
\description{
Compares the elements of two lists against each other and outputs a count table
}
\examples{
\dontrun{
set.seed(1)
l_target <- setNames(lapply(1:10, function(i) sample(letters, 5)),
                     nm = paste0("t",sprintf("\%02d", 1:10)))
l_search <- setNames(lapply(1:3, function(i) sample(letters, 3)),
                     nm = paste0("s",sprintf("\%02d", 1:3)))
res <- vcomp(query = l_search, target = l_target, rate = T)

}
}
