% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Create a chunk list from an atomic vector.}
\usage{
chunk(x, n, r)
}
\arguments{
\item{x}{atomic:}

\item{n}{integer: the length of each chunk elements}

\item{r}{logical: All of the chunks except for the last remainder are exactly 'n'.
The default value is FALSE.}
}
\description{
Create chunk list from vector.
}
\examples{
v1 <- c(1, 1, 2, NA, 3, 3, 4, 4, NA )
v2 <- factor(v1)
chunk(v1, 4)
chunk(v1, 4, TRUE)
chunk(v2, 4)

}
