% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb_utls.R
\name{kb_utls}
\alias{kb_utls}
\title{utility for kableExtra}
\usage{
kb_utls(x, kb_opt, ext_opt)
}
\arguments{
\item{x}{vector, matrix, data.frame}

\item{kb_opt}{list, options for knitr::kable,
default = list(format = 'html', align = 'c', row.names = F)}

\item{ext_opt}{list, options for kableExtra::kable_styling
default = list(bootstrap_options = 'striped', full_width = F, position = 'float_left')}
}
\description{
utility for kableExtra
}
\examples{
\dontrun{
# default parameter
kb_utls(head(iris))

# additional options of 'kable' and 'kable_styling'
kb_utls(head(iris), list(caption = 'iris', align = 'c'),
    list(bootstrap_options = 'hover'))

# no use kable_styling
kb_utls(head(iris), list(format = "pandoc", caption = "iris"), NULL)

}

}
